pyproject.toml
[tool.poetry]
name = "ModelMora"
version = "0.1.0"
description = "Neural network model management and inference serving microservice for MiraVeja"
authors = ["Jomar Júnior de Souza Pereira <jomarjunior@poli.ufrj.br>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "modelmora", from = "src" }
]

[tool.poetry.dependencies]
python = "<3.15,>=3.10"

# Web framework for REST API endpoints
fastapi = "^0.121.3"
# ASGI server for running FastAPI with async support
uvicorn = {extras = ["standard"], version = "^0.38.0"}

# gRPC framework for high-performance inference requests
grpcio = "^1.76.0"
# Tools for generating Python code from .proto files
grpcio-tools = "^1.76.0"

# Async Kafka client for event streaming
aiokafka = "^0.12.0"

# PyTorch deep learning framework with CUDA support
torch = {version = "^2.9.0", source = "torch_cuda_13"}
# Computer vision library for PyTorch
torchvision = {version = "^0.24.0", source = "torch_cuda_13"}

# YAML parser for model registry configuration
pyyaml = "^6.0.3"

# Metrics collection for Prometheus monitoring
prometheus-client = "^0.23.1"
# Structured logging for production environments
structlog = "^25.5.0"
# Load environment variables from .env files
python-dotenv = "^1.2.1"
# Dependency injection framework
miraveja-di = "^0.1.0"
miraveja-log = {extras = ["fastapi"], version = "^0.1.0"}
miraveja-authentication = "^0.1"
miraveja-protos = "^0.1.2"
transformers = "^4.57.3"
open-clip-torch = "^3.2.0"
psutil = "^7.1.3"
typing-extensions = "^4.15.0"

[[tool.poetry.source]]
name = "torch_cuda_13"
url = "https://download.pytorch.org/whl/cu130"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
# Testing framework with fixture support
pytest = "^8.2.0"
# Async test support for pytest
pytest-asyncio = "^1.3.0"
# Code coverage reporting for pytest
pytest-cov = "^7.0.0"
# gRPC testing utilities
pytest-grpc = "^0.8.0"
# Mocking library for pytest
pytest-mock = "^3.15.1"
# HTTP client for testing FastAPI endpoints
httpx = "^0.28.1"

# Code formatter for consistent style
black = "^25.11.0"
# Import statement sorter
isort = "^7.0.0"
# Python linter for code quality checks
pylint = "^4.0.3"
# Static type checker
mypy = "^1.18.2"
# Git hooks manager for automated checks
pre-commit = "^4.4.0"

# Pylint plugin for pytest compatibility
pylint-pytest = "^1.1.8"
# Pylint plugin for Pydantic model validation
pylint-pydantic = "^0.4.1"
protolint-bin = "^0.56.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["ModelMora"]

[tool.pylint.master]
ignore = ["tests", "venv", ".venv", "build", "dist", "main.py"]
jobs = 0
module-naming-style = "snake_case"
const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
function-naming-style = "snake_case"
method-naming-style = "snake_case"
attr-naming-style = "snake_case"
argument-naming-style = "snake_case"
variable-naming-style = "snake_case"
load-plugins = ["pylint_pydantic", "pylint_pytest"]
good-names = ["memory_mB"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "C0114", "C0115", "C0116", "R0917", "W0622",
    "R0903", "W0718", "R0801", "E1136", "R0913"
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/ModelMora",
    "--cov-report=html",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=80"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU"
]

[tool.coverage.run]
omit = [
    "__init__.py",
    "main.py",
]
